requirements.txt
python-telegram-bot==21.4
flask
google-generativeai